# Optimize happy path in breaker using atomics.