<!--
Request Prow to automatically lint any go code in this PR:

/lint
-->

<!-- Please include the 'why' behind your changes if no issue exists -->
## Proposed Changes

**Note:** I have to really think through more edge-cases and potentially add more test coverage but creating a PR anyway to allow more people to chew on it.

This changes the implementation of the breaker's semaphore to be based on just one atomic operation if we **know** there's capacity in the semaphore. It only relies on channel operations if there is no capacity yet to be able to park and wake up workers that want to wait for capacity in the semaphore.

It does change the semantics around error cases quite a bit in that it is not as "safe" as the old implementation is when it comes to misuse. I would say though that that doesn't really matter in our case because the error case is catastrophic anyway and we have fairly high confidence that we always pair an acquire with a release everywhere we use it.

## BenchmarkReserve

```
benchmark                                 old ns/op     new ns/op     delta
BenchmarkBreakerReserve/sequential-16     91.3          52.8          -42.17%
BenchmarkBreakerReserve/parallel-16       351           55.3          -84.25%
```

## BenchmarkMaybe

```
benchmark                                    old ns/op     new ns/op     delta
BenchmarkBreakerMaybe/1-sequential-16        88.7          25.7          -71.03%
BenchmarkBreakerMaybe/1-parallel-16          308           298           -3.25%
BenchmarkBreakerMaybe/10-sequential-16       90.1          25.4          -71.81%
BenchmarkBreakerMaybe/10-parallel-16         950           834           -12.21%
BenchmarkBreakerMaybe/100-sequential-16      89.5          25.8          -71.17%
BenchmarkBreakerMaybe/100-parallel-16        229           58.1          -74.63%
BenchmarkBreakerMaybe/1000-sequential-16     91.2          25.9          -71.60%
BenchmarkBreakerMaybe/1000-parallel-16       286           57.9          -79.76%
```